requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
einops>=0.7
lpips>=0.1.4
scikit-image>=0.21
numpy>=1.23
Pillow>=9.0
tqdm>=4.64